+++ aiohomematic/caches/dynamic.py
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2021-2025 Daniel Perna, SukramJ
 """
 Dynamic caches used at runtime by the central unit and clients.