+++ aiohomematic/model/generic/data_point.py
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2021-2025 Daniel Perna, SukramJ
 """Generic python representation of a CCU parameter."""
 
 from __future__ import annotations
@@ -16,6 +18,7 @@ from aiohomematic.const import (
     ParamsetKey,
 )
 from aiohomematic.decorators import inspector
+from aiohomematic.exceptions import ValidationException
 from aiohomematic.model import data_point as hme, device as hmd
 from aiohomematic.model.decorators import cached_slot_property
 from aiohomematic.model.support import DataPointNameData, GenericParameterType, get_data_point_name_data
@@ -101,7 +104,7 @@ class GenericDataPoint[ParameterT: GenericParameterType, InputParameterT: Generi
             return set()
         try:
             prepared_value = self._prepare_value_for_sending(value=value, do_validate=do_validate)
-        except ValueError as verr:
+        except (ValueError, ValidationException) as verr:
             _LOGGER.warning(verr)
             return set()