+++ aiohomematic/model/generic/number.py
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2021-2025 Daniel Perna, SukramJ
 """Module for data points implemented using the number category."""
 
 from __future__ import annotations
@@ -5,6 +7,7 @@ from __future__ import annotations
 from typing import cast
 
 from aiohomematic.const import DataPointCategory
+from aiohomematic.exceptions import ValidationException
 from aiohomematic.model.decorators import state_property
 from aiohomematic.model.generic.data_point import GenericDataPoint
 
@@ -30,7 +33,7 @@ class BaseDpNumber[NumberParameterT: int | float | None](GenericDataPoint[Number
             return cast(NumberParameterT, type_converter(value))
         if self._special and isinstance(value, str) and value in self._special:
             return cast(NumberParameterT, type_converter(self._special[value]))
-        raise ValueError(
+        raise ValidationException(
             f"NUMBER failed: Invalid value: {value} (min: {self._min}, max: {self._max}, special:{self._special})"
         )